app/requirements.txt
fastapi>=0.115.0                  # Core FastAPI
uvicorn[standard]>=0.30.0         # ASGI server (có standard để hỗ trợ log, reload)
sqlalchemy>=2.0.0                 # ORM
psycopg2-binary>=2.9.0            # PostgreSQL driver (sync, dễ cài trên Windows/Render)
python-dotenv>=1.0.0              # Load biến môi trường từ .env
pydantic[email]>=2.5.0            # Pydantic + email-validator (cho EmailStr)
python-jose[cryptography]>=3.3.0  # JWT (cho auth)
passlib[bcrypt]>=1.7.4            # Hash password (bcrypt)
firebase-admin>=6.0.0             # Firebase Admin SDK
python-multipart>=0.0.9           # Xử lý form-data (nếu upload file)
aiofiles>=23.0.0                  # Async file handling (nếu cần)
httpx>=0.27.0                     # HTTP client async (nếu gọi API ngoài)

# Optional nhưng hữu ích
alembic>=1.13.0                   # Migration database (rất khuyến khích khi dùng Supabase)
pydantic-settings>=2.0.0          # Config settings từ env (thay thế dotenv thủ công)